CLN: move PyFibreApplication
