+++ pyfibre/cli/pyfibre_cli.py
@@ -16,7 +16,7 @@ from pyfibre.tests.fixtures import test_shg_pl_trans_image_path
 from ..utilities import logo
 from ..version import __version__
 
-from pyfibre.cli.app.pyfibre_app import PyFibreApplication
+from pyfibre.cli.app import PyFibreApplication
 from pyfibre.shg_pl_trans.shg_pl_trans_plugin import SHGPLTransPlugin
 from pyfibre.core.core_pyfibre_plugin import CorePyFibrePlugin
